requirements.txt
PyYAML
requests
accelerate
transformers
qdrant-client
sentence-transformers
python-dateutil
python-dotenv